Migrated BitwardenRegisterConfig from backend to client module
Moved `backend.BitwardenRegisterConfig` to `client.BitwardenRegisterConfig`
of `bitwarden_passwordless` sub-package.
