+++ src/streamlit_passwordless/bitwarden_passwordless/backend.py
@@ -2,66 +2,17 @@ r"""The functions and models for interacting with the Bitwarden Passwordless bac
 
 # Standard library
 import logging
-from datetime import datetime, timedelta
-from typing import Literal, Self, TypeAlias
+from datetime import datetime
+from typing import Self
 
 # Third party
-from passwordless import PasswordlessClient, VerifiedUser
+from passwordless import VerifiedUser
 from pydantic import AnyHttpUrl
 
 # Local
-from streamlit_passwordless import common, models
+from streamlit_passwordless import models
 
 logger = logging.getLogger(__name__)
-
-
-class BitwardenRegisterConfig(models.BaseModel):
-    r"""The available passkey configuration when registering a new user.
-
-    See the `Bitwarden Passwordless`_ documentation for more info about the parameters.
-
-    .. _ Bitwarden Passwordless: https://docs.passwordless.dev/guide/api.html#register-token
-
-    Parameters
-    ----------
-    attestation : Literal['none', 'direct', 'indirect'], default 'none'
-        WebAuthn attestation conveyance preference. 'direct' and 'indirect' are exclusive to the
-        Enterprise plan of Bitwarden Passwordless. Trial & Pro plans are limited to 'none'.
-
-    authenticator_type : Literal['any', 'platform', 'cross-platform'], default 'any'
-        WebAuthn authenticator attachment modality. 'platform' refers to platform specific options
-        such as Windows Hello, FaceID or TouchID, while 'cross-platform' means roaming devices such
-        as security keys. 'any' (default) means any authenticator type is allowed.
-
-    discoverable : bool, default True
-        True allows the user to sign in without a username or alias by creating a
-        client-side discoverable credential.
-
-    user_verification : Literal['preferred', 'required', 'discouraged'], default 'preferred'
-        Set the preference for how user verification (e.g. PIN code or biometrics) works when
-        authenticating.
-
-    validity : timedelta, default timedelta(seconds=120)
-        When the registration token expires and becomes invalid defined as an offset
-        from the start of the registration process.
-
-    alias_hashing : bool, default True
-        True means that aliases for a user are hashed before they are stored in the
-        Bitwarden Passwordless database.
-    """
-
-    attestation: Literal['none', 'direct', 'indirect'] = 'none'
-    authenticator_type: Literal['any', 'platform', 'cross-platform'] = 'any'
-    discoverable: bool = True
-    user_verification: Literal['preferred', 'required', 'discouraged'] = 'preferred'
-    validity: timedelta = timedelta(seconds=120)
-    alias_hashing: bool = True
-
-    @property
-    def expires_at(self) -> datetime:
-        r"""The expiry time of the registration token in timezone UTC."""
-
-        return common.get_current_datetime() + self.validity
 
 
 class BitwardenPasswordlessVerifiedUser(models.BaseModel):
